+++ LuckParser/Models/ParseModels/Simulator/BoonSimulationItems/BoonSimulatorCleanseItem.cs
